+++ app/classes/helpers.py
@@ -32,7 +32,10 @@ class helpers:
 
     def __init__(self):
         self.crafty_root = os.path.curdir
+        self.config_dir = os.path.join(self.crafty_root, 'app', 'config')
         self.logs_dir = os.path.join(os.path.curdir, 'logs')
+
+        self.new_install_file = os.path.join(self.config_dir, 'new_install.temp')
         self.crafty_log_file = os.path.join(self.logs_dir, 'crafty.log')
         self.dbpath = os.path.join(self.crafty_root, "app", 'config', 'crafty.sqlite')
 
@@ -42,6 +45,34 @@ class helpers:
         self.passhasher = PasswordHasher()
 
         self.can_email = False
+
+    def is_fresh_install(self):
+
+        dbpath = self.get_db_path()
+
+        fresh_install = False
+
+        if not self.check_file_exists(dbpath):
+            fresh_install = True
+
+        return fresh_install
+
+    def is_setup_complete(self):
+        # if we have a new install file
+        if self.check_file_exists(self.new_install_file):
+            return False
+        else:
+            return True
+
+    def is_os_windows(self):
+        if os.name == 'nt':
+            return True
+        else:
+            return False
+
+    def get_memory(self, percentage=100):
+        mem = psutil.virtual_memory()
+        return self.human_readable_file_size(int(mem.available))
 
     def find_progam_with_server_jar(self, jar_file):
         # let's give it time to start
@@ -59,7 +90,6 @@ class helpers:
             except psutil.NoSuchProcess:
                 continue
         return False
-
 
     def get_local_ip(self):
         try:
@@ -98,6 +128,12 @@ class helpers:
         # directory already exists - non-blocking error
         except FileExistsError:
             pass
+
+    def make_new_install_file(self):
+        open(self.new_install_file, 'a').close()
+
+    def del_new_install_file(self):
+        os.remove(self.new_install_file)
 
     def check_file_exists(self, path):
         """